Moved Label, LabelText and DynLabel from button.rs to label.rs.
